+++ package.json
@@ -17,7 +17,7 @@
     "react-bootstrap-table2-paginator": "^2.1.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^4.0.3",
     "redux": "^4.1.0",
     "web-vitals": "^1.1.2"
   },
